+++ client/src/ClubDashboard.js
@@ -1,137 +1,99 @@
-import {Button, FormControl, Input, InputLabel, Container, FormHelperText,Stack, Table, TableBody, TableCell,TableContainer,TableHead,TableRow,Paper,Typography} from '@mui/material'
-import {useState, useEffect} from 'react'
-import {RequestService} from "./Services/RequestService"
-import {AddUser} from './AddUser'
-import {AddExec} from './AddExec'
+import { Button, FormControl, Input, InputLabel, Container, FormHelperText, Stack, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Paper, Typography } from '@mui/material'
+import { useState, useEffect } from 'react'
+import { RequestService } from "./Services/RequestService"
+import { AddUser } from './AddUser'
+import { AddExec } from './AddExec'
 
-function request() {
-  const requestOptions = {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    // body: JSON.stringify({
-    //   code: 'test_code',
-    //   password: 'password'
-    // })
-    body: JSON.stringify({
-      code: '8bca9603793768054af6',
-      password: 'password'
-    })
-  // const requestOptions = {
-  //   method: 'POST',
-  //   headers: { 'Content-Type': 'application/json' },
-  //   body: JSON.stringify({
-  //     ccid: 'test1',
-  //     isexec: true,
-  //     full_name: 'test man',
-  //     foip: true,
-  //     password: 'password',
-  //     clubid: 2,
-  //     token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJjY2lkIjoibWZpYXoiLCJpYXQiOjE2Mzg5MTQ2NDgsImV4cCI6MTY0MTUwNjY0OH0.Ikbba9WSkiryA9nMrtovniQYlcmSc1TWBjFNw89VXjA'
-  //  })
-  //  {
-  //   ccid: 'mfiaz',
-  //   password: 'password',
-  //   token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJjY2lkIjoibWZpYXoiLCJpYXQiOjE2Mzg5MTQ2NDgsImV4cCI6MTY0MTUwNjY0OH0.Ikbba9WSkiryA9nMrtovniQYlcmSc1TWBjFNw89VXjA'
-  // }
-    // token: JSON.stringify('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJjY2lkIjoibWZpYXoiLCJpYXQiOjE2Mzg5MTQ2NDgsImV4cCI6MTY0MTUwNjY0OH0.Ikbba9WSkiryA9nMrtovniQYlcmSc1TWBjFNw89VXjA')
-  };
-  fetch('http://localhost:8000/api/v1/email-reset/', requestOptions)
-    .then(response => response.json())
-    .then(data => console.log(data))
-    .catch(data => console.log(data));
-}
+function ClubDashboard(props) {
 
-function ClubDashboard(props){
+  const [clubName, setClubName] = useState(props.club)
+  const [ccid, setCcid] = useState("")
+  const [users, setUsers] = useState(RequestService.clubRequest(props.club, props.token))
+  const [shownUsers, setShownUsers] = useState(users)
+  const [showAddUser, setShowAddUser] = useState(false)
+  const [showAddExec, setShowAddExec] = useState(false)
 
-    const [clubName, setClubName] = useState(props.club)
-    const [ccid, setCcid] = useState("")
-    const [users, setUsers] = useState(RequestService.clubRequest(props.club,props.token))
-    const [shownUsers,setShownUsers] = useState(users)
-    const [showAddUser,setShowAddUser] = useState(false)
-    const [showAddExec,setShowAddExec] = useState(false)
+  function refresh() {
+    //get updated list of users
+    setUsers(RequestService.clubRequest(props.club, props.token))
+  }
 
-    function refresh(){
-      //get updated list of users
-      setUsers(RequestService.clubRequest(props.club,props.token))
-    }
+  function searchUsers(ccidName) {
 
-    function searchUsers(ccidName){
-    
-      console.log(ccidName)
-      ccidName = ccidName.toLowerCase()
-      if(ccidName === ""){
-        setShownUsers(users)
-      }else{
-        let newUsers = []
-        for(let user of users){
-          if (user.ccid.toLowerCase().includes(ccidName) || user.name.toLowerCase().includes(ccidName)){
-            newUsers.push(user)
-          }
+    console.log(ccidName)
+    ccidName = ccidName.toLowerCase()
+    if (ccidName === "") {
+      setShownUsers(users)
+    } else {
+      let newUsers = []
+      for (let user of users) {
+        if (user.ccid.toLowerCase().includes(ccidName) || user.name.toLowerCase().includes(ccidName)) {
+          newUsers.push(user)
         }
-        setShownUsers(newUsers)
       }
-      
+      setShownUsers(newUsers)
     }
 
-    //Table Logic
-    function createData(name, ccid, transactions) {
-        return { name, ccid, transactions };
-    }
+  }
 
-    function fetchUsers(){
-      //Get list of users from backend
+  //Table Logic
+  function createData(name, ccid, transactions) {
+    return { name, ccid, transactions };
+  }
 
-    }
-    function selectUser(ccid){
-      //Open a user's page
-      props.openUser(ccid)
-    }
+  function fetchUsers() {
+    //Get list of users from backend
 
-    function search(){
-      //Search users and return matches to show users
-      shownUsers = []
-      
-    }
+  }
+  function selectUser(ccid) {
+    //Open a user's page
+    props.openUser(ccid)
+  }
 
-    function toggleAddPerson(personType){
-      //Shows or hides add Exec and add User
-      //only one can display at a time.
+  function search() {
+    //Search users and return matches to show users
+    shownUsers = []
 
-      if(personType === "Exec"){
-        if(showAddExec){
-          //toggle closed
-          setShowAddExec(false)
-          
-        }else{
-          //toggle open
-          setShowAddExec(true)
-          if(setShowAddUser){
-            //hide add user when we open add exec
-            setShowAddUser(false)
-          }
-        }
-      }else if (personType === "customer"){
-        if(showAddUser){
-          //Toggle closed
+  }
+  function toggleAddPerson(personType) {
+    //Shows or hides add Exec and add User
+    //only one can display at a time.
+
+    if (personType === "Exec") {
+      if (showAddExec) {
+        //toggle closed
+        setShowAddExec(false)
+
+      } else {
+        //toggle open
+        setShowAddExec(true)
+        if (setShowAddUser) {
+          //hide add user when we open add exec
           setShowAddUser(false)
-        
-        }else{
-          //toggle open
-          setShowAddUser(true)
-          if(setShowAddExec){
-            //hide add user when we open add exec
-            setShowAddExec(false)
-          }
         }
       }
+    } else if (personType === "customer") {
+      if (showAddUser) {
+        //Toggle closed
+        setShowAddUser(false)
 
-
-
-
-
+      } else {
+        //toggle open
+        setShowAddUser(true)
+        if (setShowAddExec) {
+          //hide add user when we open add exec
+          setShowAddExec(false)
+        }
+      }
     }
-    
-    const table = <TableContainer component={Paper}>
+
+
+
+
+
+  }
+
+  const table = <TableContainer component={Paper}>
     <Table aria-label="simple table">
       <TableHead>
         <TableRow>
@@ -145,7 +107,7 @@ function ClubDashboard(props){
           <TableRow
             key={row.name}
             sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
-            onClick = {(e)=>{
+            onClick={(e) => {
               selectUser(row.ccid)
             }}
           >
@@ -157,40 +119,40 @@ function ClubDashboard(props){
           </TableRow>
         ))}
       </TableBody>
-      </Table>
-     </TableContainer>
+    </Table>
+  </TableContainer>
 
-    return(
-        <Stack>
-            <Stack direction = 'row' justifyContent="space-evenly">
-              <Button onClick = {request()} >Logout</Button>
-              <Button onClick = {(e)=>{selectUser(props.exec)}} >My Profile</Button>
-            </Stack>
-            
-            <Typography variant = "h1">{clubName}</Typography>
+  return (
+    <Stack>
+      <Stack direction='row' justifyContent="space-evenly">
+        <Button onClick={props.logout} >Logout</Button>
+        <Button onClick={(e) => { selectUser(props.exec) }} >My Profile</Button>
+      </Stack>
 
-            <Stack direction = "row" justifyContent = "space-between" width = "100%">
-                <Stack  direction = "row">
-                    <FormControl>
-                        <InputLabel htmlFor = "ccid">ccid or name</InputLabel>
-                        <Input id = "ccid" value = {ccid} onChange= {(e) => {setCcid(e.target.value); searchUsers(e.target.value)}} />
-                    
-                    </FormControl>
-                    <Button>Search</Button>
-                </Stack >
-                
+      <Typography variant="h1">{clubName}</Typography>
 
-                <Button onClick = {(e)=>{setShowAddUser(true)}}> Add Customer</Button>
-            </Stack>
+      <Stack direction="row" justifyContent="space-between" width="100%">
+        <Stack direction="row">
+          <FormControl>
+            <InputLabel htmlFor="ccid">ccid or name</InputLabel>
+            <Input id="ccid" value={ccid} onChange={(e) => { setCcid(e.target.value); searchUsers(e.target.value) }} />
 
-            {showAddUser && <AddUser setShowAddUser ={setShowAddUser} refresh = {refresh} />}
-            {showAddExec && <AddExec setShowAddExec ={setShowAddExec} refresh = {refresh} />}
+          </FormControl>
+          <Button>Search</Button>
+        </Stack >
 
-            {table}
 
-        </Stack>
+        <Button onClick={(e) => { setShowAddUser(true) }}> Add Customer</Button>
+      </Stack>
 
-    );
+      {showAddUser && <AddUser setShowAddUser={setShowAddUser} refresh={refresh} />}
+      {showAddExec && <AddExec setShowAddExec={setShowAddExec} refresh={refresh} />}
+
+      {table}
+
+    </Stack>
+
+  );
 
 
 }
